+++ lib/features/simple_page/simple_page.dart
@@ -1,0 +1,23 @@
+import 'package:flutter/material.dart';
+
+class SimplePage extends StatelessWidget {
+  const SimplePage({
+    required this.message,
+    super.key,
+  });
+
+  final String message;
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Text(
+        message,
+        style: Theme.of(context).textTheme.headlineSmall?.copyWith(
+              color: Theme.of(context).colorScheme.primary,
+            ),
+        textAlign: TextAlign.center,
+      ),
+    );
+  }
+}
